requirements.txt
flask
pytesseract
pillow
python-dotenv
gunicorn